requirements.txt
Jinja2==3.1.4
dateutils==0.6.12
PyYAML==6.0.1
weasyprint==62.1
